Reorganizing analysis code
